requirements.txt
streamlit
requests
pandas
textblob
plotly
wordcloud
matplotlib
python-dotenv
groq
reportlab
praw
